requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
torch
pyro-ppl